requirements.txt
crewai
crewai_tools
acp-sdk
load_dotenv
uv
nest-asyncio
colorama
smolagents
google-generativeai
langchain_huggingface
sentence-transformers